Island Making - Week 2
In this weeks lab, we learned how to make terrain. We were tasked with creating an island and populating it with grass and trees. We added textures and height differences to the island to make it more dynamic. We also messed with the lighting and the mood of the level.